[fix] Create plugin part only for `disable_on_paste.css` (#28)
There are some styles that need to be disabled on paste of large
contents into Etherpad. Those styles cannot be bundled with the other
CSS files because there won't be any way to remove them during the paste
processing, so they need to be available on separated files on Etherpad.

Implement part of https://trello.com/c/aSCjkNwv/1716.