[feat] Calculate eighths of scene (#27)
* [feat] Calculate scene length and save value on scenes

* [feat] Only apply scene length attribute if necessary

In this commit we improve the way the logic is implemented. Instead of
updating the scene length every time there is an edition, we check if
the scene length has changed. If so, we update it. With that change we
avoid collecting headings unnecessarily.

* [test] Validate scene length update on element change

* [feat] Fix bug of applying sceneLength on non-headings

This bug was found on a test on the ep_scene_navigator. The scenario was
1. Clean the script
2. Create new elements on it - necessary to create at least 1 heading
On Etherpad the second step we usually add all elements without the
outer div on a line and let the Etherpad do all job creating each
element on a new line, e.g. <div><action...<br><heading>...<br>....
Without that fix the attribute was being added on any line that a
<heading>

* [fix] Avoid updating scene length when element is hidden

This commit addresses the cases when user disable SCRIPT on EASC

* [refactor] Expose functions to calculate line dimension

* [feat] Implement cache of elements dimensions

* [feat] Minor change

Keep the EASC state to decide if it needs to run the calculation of the
scene length. When SCRIPT is disable on EASC we don't recalculate the
scene length

* [feat] Update scene length when change element type

* [fix] Remove scene length attribute on element change

This commit fixes the bug of showing a "*" in the beginning of line when
a heading is changed to a general. This fix avoids showing a scene
length class on the other script elements as well. Although, this class
has no visual effect on the element neither changes on its text.

* [refactor] Use single quotes

In this commit we only change from double quotes to single quotes

* [refactor] Address code review suggestions

* [refactor] Centralize scene length functions on ep_se

Instead of having scene length code spread across other plugins, keep it
here where we implement the logic of applying the classes