[fix] Improve performance of reset cache calculation
On this commit we improve the performance of the calculation of the
8ths. Now, we only reset the cache of the previous scene when necessary.

We had to change the logic to keep the fix of the bug 1906 (reset the
heading always when there is a scene edition).

We removed one test that does not make sense anymore. Because of the fix
of the bug 1906, we always reset the cache of the heading. It does not
matter where it was the edition.