[fix] Update duration when scene name id is provided
On this commit we fix the API that handles the scene duration. Due to
navigator UI particularities, when SCE is selected on EASC the dropdown
that changes scene duration is displayed on the SCENE item instead of on
the heading. Because of that, the scene name id is sent instead of the
heading id via API. With this fix we handle both elements ids correctly.

https://trello.com/c/sPIkPdJM/1952