[refactor] Create preAceKeyEvent to fill evt with common info (#49)
Create "preAceKeyEvent" part -- to be called before all other plugins
--, that is responsible for filling up context.evt with informations
that are useful to process an aceKeyEvent.

Currently these are the data we set:
  - isBackspace: if evt.keyCode === BACKSPACE or CTRL + H (the other way
to perform a backwards deletion on Etherpad)
  - isDelete: if evt.keyCode === DELETE
  - isRemoveKey: if key is a key that removes text (isBackspace ||
isDelete)
  - caretPosition: has flags with more advanced caret position data:
    - beginningOfLine
    - middleOfLine
    - endOfLine
  - atFirstLineOfPad
  - atLastLineOfPad

Notice: CTRL + H was not being captured as a backspace key combination,
but you can actually remove text backwards by pressing this key. It is
now included to avoid allowing users to merge lines using this key
combination.

Also add some methods to `ep_script_elements_test_helper.utils`, so they
can be used here and reused on other plugins.