[feat] Add a handler to delete script elements
In this commit we implement the "delete" feature of reformat window on
teksto. This feature allows the user to delete any element on document.
If the element is a heading or a sceneMark, we use the removeSceneMark
function of ep_script_scene_marks. If it is a regular script element, we
use the elementContentCleaner, which is implemented in this commit.