[fix] Fix error on deleting a SM through the reformat shortcut
After opening the reformat window and trying to change a "heading" to
another script element, the document got broken. To fix that, we changed
the way we define the handler for that shortcut. Then, to keep a
standard implementation, we changed the definition of the other reformat
shortcut handlers too.